Move all the insurance summary details stuff out from InsuranceSummary into its own file. Fix some styling.
